[💎 : refactor] 콘솔 에러를 대체할 에러 컴포넌트 생성 (2) - 커스텀 훅 지정 (#22)
- errorMessage state와 displayError 함수를 커스텀 훅으로 분리 : useErrorMessage
